requirements.txt
discord.py
wikipedia
beautifulsoup4
requests
lxml
selenium